+++ applications/tokenHUD.js
@@ -98,7 +98,7 @@ export async function renderHud(hud, html, token, searchText = '') {
     ? null
     : await doImageSearch(search, {
         searchType: SEARCH_TYPE.TOKEN,
-        ignoreKeywords: !worldHudSettings.includeKeywords,
+        searchOptions: { keywordSearch: worldHudSettings.includeKeywords },
       });
 
   // Merge full search, and keywords into a single array